refactor(dashboard): move link-widget and value-widget into own subdirectories
