+++ projects/element-ng/dashboard/index.ts
@@ -6,12 +6,12 @@ export * from './si-dashboard-card.component';
 export * from './si-dashboard.component';
 export * from './si-dashboard.module';
 export * from './si-dashboard.service';
-export * from './widgets/si-link-widget.component';
+export * from './widgets/si-link-widget/si-link-widget.component';
 export * from './widgets/si-list-widget/si-list-widget.component';
 export * from './widgets/si-list-widget/si-list-widget-body.component';
 export * from './widgets/si-list-widget/si-list-widget-item.component';
 export * from './widgets/si-timeline-widget/si-timeline-widget.component';
 export * from './widgets/si-timeline-widget/si-timeline-widget-body.component';
 export * from './widgets/si-timeline-widget/si-timeline-widget-item.component';
-export * from './widgets/si-value-widget-body.component';
-export * from './widgets/si-value-widget.component';
+export * from './widgets/si-value-widget/si-value-widget-body.component';
+export * from './widgets/si-value-widget/si-value-widget.component';
