+++ projects/element-ng/dashboard/widgets/si-value-widget/si-value-widget-body.component.ts
@@ -7,7 +7,7 @@ import { EntityStatusType } from '@siemens/element-ng/common';
 import { SiIconComponent, SiStatusIconComponent } from '@siemens/element-ng/icon';
 import { SiTranslatePipe, TranslatableString } from '@siemens/element-translate-ng/translate';
 
-import { SiWidgetBaseComponent } from './si-widget-base.component';
+import { SiWidgetBaseComponent } from '../si-widget-base.component';
 
 /**
  * The body of the `<si-value-widget>`. Useful for compositions.
