+++ projects/element-ng/dashboard/widgets/si-value-widget/si-value-widget.component.ts
